% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCrunch.R
\name{usun_wyniki}
\alias{usun_wyniki}
\title{Usuwa dane z bazy z danego roku.}
\usage{
usun_wyniki(ctxt, p_rok)
}
\arguments{
\item{ctxt}{Obiekt klasy DataCrunch, z którego mamy usunąć dane.}

\item{p_rok}{Rok, który chcemy usunąć.}
}
\value{
Obiekt DataCrunch z usuniętymi danymi.
}
\description{
Usuwa dane z bazy z danego roku.
}

