% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCrunch.R
\name{zaladuj_duze_dane_rok}
\alias{zaladuj_duze_dane_rok}
\title{Pobiera dane do wykresów zależne od roku.}
\usage{
zaladuj_duze_dane_rok(ctxt, rok)
}
\arguments{
\item{ctxt}{Obiekt klasy DataCrunch, w którym mamy umieścić dane.}

\item{rok}{Rok, z którego chcemy pobrać dane.}
}
\value{
Obiekt DataCrunch z umieszczonymi nowymi danymi.
}
\description{
Dane pobierane w tej funkcji nie zmieniają się, lecz co roku dodawane są
do nich nowe dane. Aby pobrać dane z danego roku, wywołaj funkcję z odpowiednim
parametrem.
}

