% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCrunch.R
\name{zaladuj_male_dane}
\alias{zaladuj_male_dane}
\title{Pobiera dane pomocnicze niezależne od roku.}
\usage{
zaladuj_male_dane(ctxt)
}
\arguments{
\item{ctxt}{Obiekt klasy DataCrunch, w którym mamy umieścić dane.}
}
\value{
Obiekt DataCrunch z umieszczonymi nowymi danymi.
}
\description{
Część przetwarzanych danych pełni funkcję pomocniczą w przetwarzaniu
wyników egzaminów. Ta funkcja wczytuje je.
}
\details{
Dane pobierane w tej funkcji mogą zmieniać się co roku. Wywołanie jej
zastąpi poprzednie dane w dazie danych - aby zaktualizować dane, wystarczy
więc wywołać tę funkcję raz.
}

