% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCrunch.R
\name{zaladuj_nowe_wyniki}
\alias{zaladuj_nowe_wyniki}
\title{Pozwala umieścić w danych wyniki danego egzaminu.}
\usage{
zaladuj_nowe_wyniki(ctxt, p_rok, t_rodzaj, t_czesc, l_dane)
}
\arguments{
\item{ctxt}{Obiekt klasy DataCrunch, w którym mamy umieścić dane.}

\item{p_rok}{Rok egzaminu.}

\item{t_rodzaj}{Rodzaj egzaminu.}

\item{t_czesc}{Część egzaminu.}

\item{l_dane}{Dane pobrane przez pakiet ZPD funkcją pobierz_wyniki_egzaminu(..., czy_ewd = TRUE).}
}
\value{
Obiekt danych DataCrunch z dodanymi danymi, jeśli dane zostały poprawnie pobrane
i wprowadzone do bazy. Niezmieniony obiekt, jeśli dane egzaminu już zostały wprowadzone.
}
\description{
UWAGA: dane, które w bazie już się znajdują zostaną zignorowane. Jeśli chcesz wprowadzić
je ponownie, użyj funkcji usun_wyniki.
UWAGA: zalecane jest chronologiczne wprowadzanie danych.
Patrz vignette(package="ZPD.dataCrunch", topic="wyniki_po_egz.
}

